+++ script/migrations/occurrence/2.6.0_to_2.7.0.sql
@@ -11,3 +11,9 @@ WHERE typestatus IS NOT NULL AND typestatus <> '' AND occ_raw.auto_id = occ.auto
 UPDATE occurrence occ SET typeStatus = occ_raw.typeStatus
 FROM occurrence_raw occ_raw
 WHERE occ_raw.auto_id = occ.auto_id;
+
+--WSG84 shifted latitude/longitude (result of ST_Shift_Longitude)
+SELECT AddGeometryColumn('public','occurrence','the_shifted_geom', 4326, 'POINT', 2 );
+SELECT AddGeometryColumn('buffer','occurrence','the_shifted_geom', 4326, 'POINT', 2 );
+
+UPDATE occurrence SET the_shifted_geom = ST_Shift_Longitude(the_geom) WHERE the_geom IS NOT NULL;